refactor(core): modularize native DB and pseudo-terminal logic
- Removed inline native db and pseudo-terminal modules from `nx` package.
- Moved `pseudo_terminal` modules to `nx_pty` crate.
- Moved DB connection handling into `nx_db` crate under a new `napi` module.
- Updated dependencies and adjusted imports for modular separation.
- Dropped unused dependencies (`tokio`, `crossbeam-channel`, etc.) for cleanup.
