+++ crates/nx_db/Cargo.toml
@@ -5,10 +5,16 @@ edition.workspace = true
 license.workspace = true
 
 [dependencies]
-rusqlite = { workspace = true }
+napi = { workspace = true }
+napi-derive = { workspace = true }
+nx_logger = { workspace = true }
+nx_utils = { workspace = true }
 anyhow = { workspace = true }
 tracing = { workspace = true }
 fs4 = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+rusqlite = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
