+++ crates/nx_db/src/napi/mod.rs
@@ -1,9 +1,10 @@
+use crate::NxDbConnection;
 use napi::bindgen_prelude::External;
+use napi_derive::*;
+use nx_logger::enable_logger;
 use std::fs::create_dir_all;
 use std::path::PathBuf;
 
-use nx_db::connection::NxDbConnection;
-use nx_logger::enable_logger;
 use nx_utils::machine_id::get_machine_id;
 
 #[napi]
@@ -17,11 +18,11 @@ pub fn connect_to_nx_db(
     create_dir_all(&cache_dir_buf)?;
 
     let machine_id = get_machine_id();
-    let connection = nx_db::connect_to_nx_db(cache_dir, nx_version, db_name, machine_id)?;
+    let connection = crate::open_db_connection(cache_dir, nx_version, db_name, machine_id)?;
     Ok(External::new(connection))
 }
 
 #[napi]
 pub fn close_db_connection(mut connection: External<NxDbConnection>) -> anyhow::Result<()> {
-    nx_db::close_db_connection(std::mem::take(connection.as_mut()))
+    crate::close_db_connection(std::mem::take(connection.as_mut()))
 }
