+++ crates/nx_pty/Cargo.toml
@@ -5,13 +5,16 @@ license.workspace = true
 version.workspace = true
 
 [dependencies]
+nx_logger = { workspace = true }
 napi = { workspace = true }
 napi-derive = { workspace = true }
-portable-pty = { workspace = true }
 anyhow = { workspace = true }
 crossbeam-channel = { workspace = true }
-crossterm = { workspace = true }
 tracing = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+portable-pty = { workspace = true }
+crossterm = { workspace = true }
 
 [target.'cfg(all(not(windows), not(target_family = "wasm")))'.dependencies]
 mio = { workspace = true }
