+++ crates/nx_pty/src/lib.rs
@@ -1,1 +1,4 @@
+pub mod napi;
+
+#[cfg(not(target_arch = "wasm32"))]
 pub mod pseudo_terminal;
