+++ packages/nx/src/native/mod.rs
@@ -8,9 +8,8 @@ mod types;
 mod utils;
 pub mod workspace;
 
-#[cfg(not(target_arch = "wasm32"))]
-pub mod db;
-#[cfg(not(target_arch = "wasm32"))]
-pub mod pseudo_terminal;
+
+
+
 #[cfg(not(target_arch = "wasm32"))]
 pub mod watch;
